requirements.txt
Flask==3.0.0
openpyxl==3.1.2
Werkzeug==3.0.1
Pillow==10.4.0
